Added New Assets to Existing Code
What is changed:

Made new Atlas with new sprite assets, added to Game.ts (core folder)

*Will rename classes accordingly after this level is finished*

Barrel.ts is for Pickups - so only changed the Sprite stringname
-same for Enemy.ts, except there is no vertical drag, just spawns at
random y-position and speeds across the scene horizontally
-same for Ship, changed sprite stringname in the constructor

Menu, Game, Over.ts - reworded stuff, and add _playerShot reference - of
the playerShot class

*MOST IMPORTANT: playerShot Class
-tried to make the Blast2 sprite a gameobject in this class
-I want to instantiate it in Game Scene (or any scene with addChild
since the Player Ship will be in 3 levels = 3 scenes)
-positioned in front of the Player Space Ship, and is accelerated by a
speed variable across screen (reference this playerShot gameObject to
Collision Class or use the Collision methods from a reference of type
Collision Class in playerShot Class to look for collision with enemies -
upon hit, enemy is reset, adds to player score -- or might even have to
make a new collision class just for player shot)
-in order to shoot a playerShot, there is a click event handler in the
Game Scene, which calls a method to instantiate a reference of type
playerShot and addChild to the Game Scene/to the createJS.Stage
-last but not least, once playerShot is out of the canvas/game scene
(beyond 640 width), destroy the playerShot objects, or basically call
the update method of the playerShot reference in the scene to
removeChild, or call the update method of the Scene to execute that
condition...

What needs to be done, at least for this first Level:

-Shoot projectiles by both the player (via right click or some other
button mapping) and by the enemy if possible (at a fix firing rate if
Create or Easel JS has a Time.time type of class/functionality to
program the continuous shots

-Add new audio for enemy shots/hit detection, player shots (already have
hit detection sound), soundtrack, pickups, etc.

-Add Pickups-
*Box is the Fusion Core - need 15 to get through the level
*Small Triangle thing is could be new weapon power up - that shoots a
longer laser that goes through enemies in a clear path (doesn't
disappear upon collision, keep destroying oncoming enemies) -- for now
it adds to the score as valuable loose elements (uranium-fusion) adrift
in space.
*The other small pickup can be weapon mods and such, for now, they add
to your score, or could be score multipliers treated as getting more
valuable elements
*The Cross could give player 1 Extra Life